+++ app/old/PostRepositoryFileImpl.kt
@@ -1,69 +1,87 @@
-package ru.netology.nmedia.data.repository
+package ru.netology.nmedia.data.repository.old
 
-import android.app.Application
 import android.content.Context
+import android.util.Log
 import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import ru.netology.nmedia.R
 import ru.netology.nmedia.domain.repository.PostRepository
 import ru.netology.nmedia.dto.NON_EXISTING_POST_ID
 import ru.netology.nmedia.dto.Post
 import java.util.*
 
-class PostRepositorySharedPrefsImpl(context: Context) : PostRepository {
+class PostRepositoryFileImpl(private val context: Context) :
+    PostRepository {
     private val gson = Gson()
-    private val prefs = context.getSharedPreferences("repo", Context.MODE_PRIVATE)
     private val type = TypeToken.getParameterized(List::class.java, Post::class.java).type
-    private val key = "posts"
+    private val filename = "posts.json"
     private var nextId = 0L
     private var posts = emptyList<Post>()
-    private val data = MutableLiveData(posts)
+
+    private val _data = MutableLiveData(emptyList<Post>())
+
+    val data: LiveData<List<Post>>
+        get() = _data
+
 
     private val defaultPosts = List(5) {
         Post(
             id = ++nextId,
-            author = "Нетология ${nextId.toLong()}. Университет интернет-профессий будущего",
+            author = "Нетология ${nextId}. Университет интернет-профессий будущего",
             content = "Знаний хватит на всех: на следующей неделе разбираемся с чем-то еще",
             published = "15 окт в 15:28",
             likedByMe = false,
             likes = 999,
-            shares = 2999,
+            shares = 2999
         )
     }.onEach { if (it.id == 2L) it.video = "https://www.youtube.com/watch?v=WhWc3b3KhnY" }
 
     init {
-        prefs.getString(key, null).let {
-            posts = if (it != null) {
-                Toast.makeText(context, "Loaded posts from sharedPrefs!", Toast.LENGTH_LONG).show()
-                gson.fromJson(it, type)
-            } else {
-                Toast.makeText(context, "Loaded default DEMO posts!", Toast.LENGTH_LONG).show()
-                defaultPosts
+        try {
+            val file = context.filesDir.resolve(filename)
+            file.bufferedReader().use {
+                posts = gson.fromJson(it, type)
             }
-
-            nextId = posts.maxOf { it.id }.plus(1)
-            data.value = posts
+        } catch (e: Exception) {
+            Log.w(
+                "Settings", "Error while loading file [${filename}]: ${e.message}" +
+                        "\nLoading default settings from DEMO posts!"
+            )
+            Toast.makeText(context, "Loaded default DEMO posts!", Toast.LENGTH_LONG).show()
+            posts = defaultPosts
         }
+
+        nextId = posts.maxOfOrNull { it.id }?.plus(1) ?: 0
+        _data.value = posts
+
+        sync()
     }
+
 
     private fun sync() {
-        with(prefs.edit()) {
-            putString(key, gson.toJson(posts))
-            commit()
+        try {
+            val file = context.filesDir.resolve(filename)
+            file.bufferedWriter().use {
+                it.write(gson.toJson(posts))
+            }
+        } catch (e: Exception) {
+            Log.e("Settings", "Error while saving file [${filename}]: ${e.message}")
+            Toast.makeText(
+                context,
+                context.getString(R.string.settings_save_error_message),
+                Toast.LENGTH_LONG
+            )
+                .show()
         }
-
     }
 
 
-
-
-    //private val data = MutableLiveData(posts)
-
-    override fun getAll(): LiveData<List<Post>> {
-        return Transformations.map(data) { input -> input.sortedByDescending { it.id } }
+    override fun getAll() {
+        _data.value = Transformations.map(data) { input -> input.sortedByDescending { it.id } }.value
     }
 
     override fun likeById(id: Long) {
@@ -74,7 +92,7 @@ class PostRepositorySharedPrefsImpl(context: Context) : PostRepository {
             else it
         }
 
-        data.value = posts
+        _data.value = posts
         sync()
     }
 
@@ -84,13 +102,13 @@ class PostRepositorySharedPrefsImpl(context: Context) : PostRepository {
             else it
         }
 
-        data.value = posts
+        _data.value = posts
         sync()
     }
 
     override fun removeById(id: Long) {
         posts = posts.filter { it.id != id }
-        data.value = posts
+        _data.value = posts
         sync()
     }
 
@@ -117,8 +135,11 @@ class PostRepositorySharedPrefsImpl(context: Context) : PostRepository {
                     it
             }
         }
-        data.value = posts
+        _data.value = posts
         sync()
 
     }
+
+
+
 }
