+++ app/old/PostRepositorySQLiteRoomImpl.kt
@@ -1,4 +1,4 @@
-package ru.netology.nmedia.data.repository
+package ru.netology.nmedia.data.repository.old
 
 import android.content.Context
 import androidx.lifecycle.LiveData
@@ -12,16 +12,21 @@ import ru.netology.nmedia.dto.PostEntity
 
 class PostRepositorySQLiteRoomImpl(context: Context) : PostRepository {
     private val dao: PostDao = AppDb.getInstance(context).postDao()
-    private val data = MutableLiveData(emptyList<Post>())
+    private val _data = MutableLiveData(emptyList<Post>())
+
+    val data: LiveData<List<Post>>
+        get() = _data
 
     init {
-        data.value = getAll().value
+        //data.value = getAll().value
     }
 
-    override fun getAll(): LiveData<List<Post>> = Transformations.map(dao.getAll()) { list ->
-        list.map {
-            PostEntity.toDto(it)
-        }
+    override fun getAll() {
+        _data.value = Transformations.map(dao.getAll()) { list ->
+            list.map {
+                PostEntity.toDto(it)
+            }
+        }.value
     }
 
     override fun likeById(id: Long) {
@@ -39,4 +44,6 @@ class PostRepositorySQLiteRoomImpl(context: Context) : PostRepository {
     override fun addOrEditPost(post: Post) {
         dao.addOrEditPost(PostEntity.fromDto(post))
     }
+
+
 }
